Workaround for Qt6 QFontMetrics::height()/descent() bugs
descent() wrongly returns negative values (should be positive as per docs),
height() according to docs is a sum of ascent() and descent(), so now it
returns a value that has no sense whatsoever.

The workaround is to provide our own height() function that returns sum
of absolute values of ascent() and descent() (making sure the code won't
break if Qt again changes the API in the future).